refactor: Adjust structure and delete useless codes
