--- internal/core/mcpclient/transport/interface.go
@@ -1,25 +1,0 @@
-package transport
-
-import (
-	"context"
-
-	"github.com/mcp-ecosystem/mcp-gateway/pkg/mcp"
-)
-
-type Interface interface {
-	// Start the connection. Start should only be called once.
-	Start(ctx context.Context) error
-
-	// SendRequest sends a json RPC request and returns the response synchronously.
-	SendRequest(ctx context.Context, request mcp.JSONRPCRequest) (*mcp.JSONRPCResponse, error)
-
-	// SendNotification sends a json RPC Notification to the server.
-	SendNotification(ctx context.Context, notification mcp.JSONRPCNotification) error
-
-	// SetNotificationHandler sets the handler for notifications.
-	// Any notification before the handler is set will be discarded.
-	SetNotificationHandler(handler func(notification mcp.JSONRPCNotification))
-
-	// Close the connection.
-	Close() error
-}
