+++ internal/core/tool.go
@@ -2,8 +2,14 @@ package core
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/gin-gonic/gin"
+	"github.com/mark3labs/mcp-go/client"
+	"github.com/mark3labs/mcp-go/client/transport"
+	mcpgo "github.com/mark3labs/mcp-go/mcp"
+	"github.com/mcp-ecosystem/mcp-gateway/internal/mcp/session"
 	"io"
 	"mime/multipart"
 	"net/http"
@@ -59,6 +65,28 @@ func prepareTemplateContextForMCPBackend(args map[string]any, request *http.Requ
 	}
 
 	return tmplCtx, nil
+}
+
+func preprocessArgs(args map[string]any) map[string]any {
+	processed := make(map[string]any)
+
+	for k, v := range args {
+		switch val := v.(type) {
+		case []any:
+			ss, _ := json.Marshal(val)
+			processed[k] = string(ss)
+		case float64:
+			// If the float64 equals its integer conversion, it's an integer
+			if val == float64(int64(val)) {
+				processed[k] = int64(val)
+			} else {
+				processed[k] = val
+			}
+		default:
+			processed[k] = v
+		}
+	}
+	return processed
 }
 
 // prepareRequest prepares the HTTP request with templates and arguments
@@ -202,24 +230,315 @@ func (s *Server) executeHTTPTool(tool *config.ToolConfig, args map[string]any, r
 	return callToolResult, nil
 }
 
-func preprocessArgs(args map[string]any) map[string]any {
-	processed := make(map[string]any)
+func (s *Server) executeStdioTool(
+	c *gin.Context,
+	tool *config.MCPServerConfig,
+	args map[string]any,
+	request *http.Request,
+	params mcp.CallToolParams,
+) (*mcp.CallToolResult, error) {
+	// Prepare template context
+	tmplCtx, err := prepareTemplateContextForMCPBackend(args, request)
+	if err != nil {
+		return nil, err
+	}
 
-	for k, v := range args {
-		switch val := v.(type) {
-		case []any:
-			ss, _ := json.Marshal(val)
-			processed[k] = string(ss)
-		case float64:
-			// If the float64 equals its integer conversion, it's an integer
-			if val == float64(int64(val)) {
-				processed[k] = int64(val)
-			} else {
-				processed[k] = val
+	renderedClientEnv := make(map[string]string)
+	for k, v := range tool.Env {
+		rendered, err := renderTemplate(v, tmplCtx)
+		if err != nil {
+			return nil, err
+		}
+		renderedClientEnv[k] = rendered
+	}
+
+	toolCallRequestParams := make(map[string]interface{})
+	if err := json.Unmarshal(params.Arguments, &toolCallRequestParams); err != nil {
+		return nil, err
+	}
+
+	// Use mcp-go client
+	command := tool.Command
+	cmdArgs := tool.Args
+	stdioClientEnv := mcp.CoverToStdioClientEnv(renderedClientEnv)
+
+	// Create stdio transport
+	stdioTransport := transport.NewStdio(command, stdioClientEnv, cmdArgs...)
+
+	// Start the transport
+	if err := stdioTransport.Start(c.Request.Context()); err != nil {
+		return nil, fmt.Errorf("failed to start stdio transport: %w", err)
+	}
+
+	// Create client
+	mcpClient := client.NewClient(stdioTransport)
+	defer mcpClient.Close()
+
+	// Initialize client
+	initRequest := mcpgo.InitializeRequest{}
+	initRequest.Params.ProtocolVersion = mcpgo.LATEST_PROTOCOL_VERSION
+	initRequest.Params.ClientInfo = mcpgo.Implementation{
+		Name:    "mcp-gateway",
+		Version: "0.1.0",
+	}
+
+	_, err = mcpClient.Initialize(c.Request.Context(), initRequest)
+	if err != nil {
+		return nil, fmt.Errorf("failed to initialize stdio client: %w", err)
+	}
+
+	// Call tool
+	callRequest := mcpgo.CallToolRequest{}
+	callRequest.Params.Name = params.Name
+
+	// Convert parameters to mcp-go format
+	callRequest.Params.Arguments = toolCallRequestParams
+
+	mcpgoResult, err := mcpClient.CallTool(c.Request.Context(), callRequest)
+	if err != nil {
+		return nil, fmt.Errorf("failed to call tool: %w", err)
+	}
+
+	// Convert mcp-go result to local mcp format
+	result := &mcp.CallToolResult{
+		IsError: mcpgoResult.IsError,
+	}
+
+	// Process content items
+	if len(mcpgoResult.Content) > 0 {
+		var validContents []mcp.Content
+
+		for _, content := range mcpgoResult.Content {
+			// Skip null content
+			if content == nil {
+				continue
 			}
-		default:
-			processed[k] = v
+
+			// Try to get content type
+			contentType := ""
+			switch c := content.(type) {
+			case *mcpgo.TextContent:
+				contentType = "text"
+				validContents = append(validContents, &mcp.TextContent{
+					Type: "text",
+					Text: c.Text,
+				})
+			case *mcpgo.ImageContent:
+				contentType = "image"
+				validContents = append(validContents, &mcp.ImageContent{
+					Type:     "image",
+					Data:     c.Data,
+					MimeType: c.MIMEType,
+				})
+			case *mcpgo.AudioContent:
+				contentType = "audio"
+				validContents = append(validContents, &mcp.AudioContent{
+					Type:     "audio",
+					Data:     c.Data,
+					MimeType: c.MIMEType,
+				})
+			default:
+				// Try to parse from raw content
+				rawContent, err := json.Marshal(content)
+				if err == nil {
+					var contentMap map[string]interface{}
+					if json.Unmarshal(rawContent, &contentMap) == nil {
+						if typ, ok := contentMap["type"].(string); ok {
+							contentType = typ
+
+							switch contentType {
+							case "text":
+								if text, ok := contentMap["text"].(string); ok {
+									validContents = append(validContents, &mcp.TextContent{
+										Type: "text",
+										Text: text,
+									})
+								}
+							case "image":
+								data, _ := contentMap["data"].(string)
+								mimeType, _ := contentMap["mimeType"].(string)
+								validContents = append(validContents, &mcp.ImageContent{
+									Type:     "image",
+									Data:     data,
+									MimeType: mimeType,
+								})
+							case "audio":
+								data, _ := contentMap["data"].(string)
+								mimeType, _ := contentMap["mimeType"].(string)
+								validContents = append(validContents, &mcp.AudioContent{
+									Type:     "audio",
+									Data:     data,
+									MimeType: mimeType,
+								})
+							}
+						}
+					}
+				}
+			}
+		}
+
+		if len(validContents) > 0 {
+			result.Content = validContents
 		}
 	}
-	return processed
+
+	return result, nil
+}
+
+func (s *Server) fetchHTTPToolList(conn session.Connection) ([]mcp.ToolSchema, error) {
+	// Get http tools for this prefix
+	tools, ok := s.state.prefixToTools[conn.Meta().Prefix]
+	if !ok {
+		tools = []mcp.ToolSchema{} // Return empty list if prefix not found
+	}
+
+	return tools, nil
+}
+
+func (s *Server) fetchStdioToolList(ctx context.Context, conn session.Connection) ([]mcp.ToolSchema, error) {
+	// Get stdio tools for this prefix
+	stdioCfg, ok := s.state.prefixToMCPServerConfig[conn.Meta().Prefix]
+	if !ok {
+		return []mcp.ToolSchema{}, nil
+	}
+
+	// Create stdio transport with the command and arguments
+	command := stdioCfg.Command
+	args := stdioCfg.Args
+	stdioClientEnv := mcp.CoverToStdioClientEnv(stdioCfg.Env)
+
+	// Create mcp-go stdio transport
+	stdioTransport := transport.NewStdio(command, stdioClientEnv, args...)
+
+	// Start the transport
+	if err := stdioTransport.Start(ctx); err != nil {
+		return []mcp.ToolSchema{}, fmt.Errorf("failed to start stdio transport: %w", err)
+	}
+
+	// Create client with the transport
+	c := client.NewClient(stdioTransport)
+	defer c.Close()
+
+	// Initialize the client
+	initRequest := mcpgo.InitializeRequest{}
+	initRequest.Params.ProtocolVersion = mcpgo.LATEST_PROTOCOL_VERSION
+	initRequest.Params.ClientInfo = mcpgo.Implementation{
+		Name:    "mcp-gateway",
+		Version: "0.1.0",
+	}
+
+	_, err := c.Initialize(ctx, initRequest)
+	if err != nil {
+		return []mcp.ToolSchema{}, fmt.Errorf("failed to initialize: %w", err)
+	}
+
+	// List available tools
+	toolsRequest := mcpgo.ListToolsRequest{}
+	toolsResult, err := c.ListTools(ctx, toolsRequest)
+	if err != nil {
+		return []mcp.ToolSchema{}, fmt.Errorf("failed to list tools: %w", err)
+	}
+
+	// Convert from mcpgo.Tool to mcp.ToolSchema
+	tools := make([]mcp.ToolSchema, len(toolsResult.Tools))
+	for i, schema := range toolsResult.Tools {
+		// Create local mcp package ToolInputSchema
+		inputSchema := mcp.ToolInputSchema{
+			Type:       "object",
+			Properties: make(map[string]any),
+		}
+
+		// Convert mcpgo InputSchema to local mcp format
+		rawSchema, err := json.Marshal(schema.InputSchema)
+		if err == nil {
+			// Parse schema properties
+			var schemaMap map[string]interface{}
+			if err := json.Unmarshal(rawSchema, &schemaMap); err == nil {
+				if properties, ok := schemaMap["properties"].(map[string]interface{}); ok {
+					inputSchema.Properties = properties
+				}
+				if typ, ok := schemaMap["type"].(string); ok {
+					inputSchema.Type = typ
+				}
+				if required, ok := schemaMap["required"].([]interface{}); ok {
+					reqStrings := make([]string, len(required))
+					for j, r := range required {
+						if rStr, ok := r.(string); ok {
+							reqStrings[j] = rStr
+						}
+					}
+					inputSchema.Required = reqStrings
+				}
+			}
+		}
+
+		tools[i] = mcp.ToolSchema{
+			Name:        schema.Name,
+			Description: schema.Description,
+			InputSchema: inputSchema,
+		}
+	}
+
+	return tools, nil
+}
+
+func (s *Server) invokeHTTPTool(c *gin.Context, req mcp.JSONRPCRequest, conn session.Connection, params mcp.CallToolParams) *mcp.CallToolResult {
+	// Find the tool in the precomputed map
+	tool, exists := s.state.toolMap[params.Name]
+	if !exists {
+		errMsg := "Tool not found"
+		s.sendProtocolError(c, req.Id, errMsg, http.StatusNotFound, mcp.ErrorCodeMethodNotFound)
+		return nil
+	}
+
+	// Convert arguments to map[string]any
+	var args map[string]any
+	if err := json.Unmarshal(params.Arguments, &args); err != nil {
+		errMsg := "Invalid tool arguments"
+		s.sendProtocolError(c, req.Id, errMsg, http.StatusBadRequest, mcp.ErrorCodeInvalidParams)
+		return nil
+	}
+
+	// Get server configuration
+	serverCfg, ok := s.state.prefixToServerConfig[conn.Meta().Prefix]
+	if !ok {
+		errMsg := "Server configuration not found"
+		s.sendProtocolError(c, req.Id, errMsg, http.StatusInternalServerError, mcp.ErrorCodeInternalError)
+		return nil
+	}
+
+	// Execute the tool
+	result, err := s.executeHTTPTool(tool, args, c.Request, serverCfg.Config)
+	if err != nil {
+		s.sendToolExecutionError(c, conn, req, err, true)
+		return nil
+	}
+
+	return result
+}
+
+func (s *Server) invokeStdioTool(c *gin.Context, req mcp.JSONRPCRequest, conn session.Connection, params mcp.CallToolParams) *mcp.CallToolResult {
+	// Get stdio tools for this prefix
+	stdioCfg, ok := s.state.prefixToMCPServerConfig[conn.Meta().Prefix]
+	if !ok {
+		errMsg := "Server configuration not found"
+		s.sendProtocolError(c, req.Id, errMsg, http.StatusNotFound, mcp.ErrorCodeMethodNotFound)
+		return nil
+	}
+
+	// Convert arguments to map[string]any
+	var args map[string]any
+	if err := json.Unmarshal(params.Arguments, &args); err != nil {
+		errMsg := "Invalid tool arguments"
+		s.sendProtocolError(c, req.Id, errMsg, http.StatusBadRequest, mcp.ErrorCodeInvalidParams)
+		return nil
+	}
+
+	result, err := s.executeStdioTool(c, &stdioCfg, args, c.Request, params)
+	if err != nil {
+		s.sendToolExecutionError(c, conn, req, err, true)
+		return nil
+	}
+
+	return result
 }
